fix: multiple borrows with stack pointers
Having stack pointers stored on the stack was causing a multiple-
borrow error. Since checking the id of a region requires borrowing.
When cloning the stack, ids stored in pointers are compared against
ids from regions, resulting in a query for the stack id, borrowing the
stack against.

The fix just stores the stack id temporarily, as is done for other
regions.